Included Update to EMG Filter
Added median filter at end of EG pre-processing pipeline, with a conditional statement on whether it is being applied to MVC or Gait, which will adjust the smoothing window length accordingly. Reason is to reduce capturing high frequency noise as the peak activation. Data has been updated and processed for SS10.