Slight Updates in EMG load, plotting, and hardware indices
Made loading EMG more robust to not always remove the last channel, because there are situations where a recording is done without that channel active, and there for the -1 strategy was removing the final muscle.

Originally I was plotting cycleTable instead of matchedCycleTable by mistake, fixed visuals as well.

From plotting the data, there were instances where it seemed like there was a lag or shift, I didn't quite understand the complexity in this original code, and adjusted to what was needed, which seemed to resolve the issue of lag.